Refactor code into function groups
