+++ .gitignore
@@ -21,3 +21,4 @@ target
 #.idea/
 
 Cargo.lock
+rustc-ice-*
